+++ CHANGELOG.md
@@ -10,6 +10,10 @@ project uses [Semantic Versioning].
 - The `on_error` parameter for `Importer.sync_package`, `Package.sync_resource`,
   and `Resource.sync_view` now defaults to `OnError.reraise`.
 
+### Added
+
+- Support for pagination during package search
+
 ### Fixed
 
 - `Importer.sync_package` did not find existing private packages.